More memory
Currently, the buildpack does not have access to the $MEMORY_LIMIT
environment variable.  Because of this, the JRE is not being
configured to limit itself to the amount of memory available.  This
change bumps the amount of memory to a size large enough that the
default JVM memory limits won't cause any problems.

[#49694799]