No classes in Groovy application
Due to recent changes to what constitutes a Groovy application, the
existence of classes in the sample application meant that it would not
run.  This change re-jiggers the application such that it no longer
needs classes to run (using a GroovyClassLoader rather than a
URLClassLoader) allowing it to be started properly by the buildpack.