Handle FakeMongoDbFactory correctly
Previously, if the MongoDB URL was requested, but the MongoDB had not
been reconfigured it would result in an NPE.  During testing, we never
called for the URL in that case, but either way it's not a good
result.  This change protects against that possibility and returns an
dummy string for the URL in the case where the MongoDB has not been
reconfigured.

[resolves #9]