Better OOM diagnostics
As part of the diagnosis of #77279898, it was important to be able to
create an OOM condition above 2GB.  Previously, the applications did
not have a way to trigger an OOM at will (only on startup) and there
was no indication of how much memory was being consumed while the OOM
as happening.  This change adds a /out-of-memory endpoint and logging
to demonstrate how much memory is being used during exhaustion.

[#77279898]