Remove provided dependencies
Previously the list of dependencies for the web-application project
included both the servlet-api and components of embedded Tomcat.
These dependencies caused issues when starting the application in tc-
server.

This change updates the projects (specifically, the core and spring-
common projects) to make those dependencies provided so that they are
not included in downstream application projects.